+++ src/DPMatrix.h
@@ -4,8 +4,8 @@ class DPMatrix {
     public:
         DPMatrix(const int &s1_l, const int &s2_l);
         ~DPMatrix();
-        int get_pos(const int &i, const int &j, const int &k, const int &l);
-        void put_pos(const int &i, const int &j, const int &k, const int &l, const int &val);
+        float get_pos(const int &i, const int &j, const int &k, const int &l);
+        void put_pos(const int &i, const int &j, const int &k, const int &l, const float &val);
         long long int get_total_size() const;
         bool check_border(const int &i, const int &j, const int &k, const int &l) const;
 
@@ -13,7 +13,7 @@ class DPMatrix {
         int calc_delta(int i, int j, int k, int l) const;
         long long int calc_total_size(long long int s1, long long int s2) const;
 
-        int *dp_matrix;
+        float *dp_matrix;
         const int s1_l;
         const int s2_l;
 };
