+++ src/DPMatrix_GPU.h
@@ -13,9 +13,9 @@ struct sequences {
 long long int dp_matrix_calc_total_size(long long int s1, long long int s2);
 #ifdef __CUDACC__
 __device__ __host__ int dp_matrix_calc_delta(int i, int j, int k, int l, sequences* seq_ctx);
-__device__ __host__ int dp_matrix_get_pos(int *dp_matrix, const int &i, const int &j, const int &k, const int &l, sequences* seq_ctx);
-__device__ __host__ void dp_matrix_put_pos(int *dp_matrix, const int &i, const int &j, const int &k, const int &l, const int &val, sequences* seq_ctx);
-__host__ int dp_matrix_get_val(int *dp_matrix, const int &i, const int &j, const int &k, const int &l, sequences* seq_ctx);
+__device__ __host__ float dp_matrix_get_pos(float *dp_matrix, const int &i, const int &j, const int &k, const int &l, sequences* seq_ctx);
+__device__ __host__ void dp_matrix_put_pos(float *dp_matrix, const int &i, const int &j, const int &k, const int &l, const float &val, sequences* seq_ctx);
+__host__ float dp_matrix_get_val(float *dp_matrix, const int &i, const int &j, const int &k, const int &l, sequences* seq_ctx);
 
 __device__ __host__ inline bool dp_matrix_check_border(const int &i, const int &j, const int &k, const int &l, sequences* seq_ctx)
 {
